Change file names

+++ database/migrations/2021_01_11_074435_create_posts_table.php
@@ -18,8 +18,8 @@ class CreatePostsTable extends Migration
             $table->string('title')->unique();
             $table->string('description');
             $table->integer('status')->default(1);
-            $table->integer('created_user_id');
-            $table->integer('updated_user_id');
+            $table->unsignedBigInteger('created_user_id');
+            $table->unsignedBigInteger('updated_user_id');
             $table->integer('deleted_user_id')->nullable();
             $table->datetime('created_at');
             $table->datetime('updated_at');
